// citas-reservacion-sistema-frontend/src/types/UserData.ts
export interface UserData {
  dni: string;
  nombre: string;
  apellidoPaterno: string;
  apellidoMaterno: string;
  fechaNacimiento: string;
  telefono: string;
  email: string;
  genero: string;
  departamento: string;
  provincia: string;
  distrito: string;
  direccion: string;
  password: string;
}